taking OrderedEvent and OrderedL2ToL1Message objects from starknet_api and adding them to FunctionInvokation object. Added ThinStateDiff to InvokeTransactionTrace, this is wip and the ThinStateDiff from currently does not have the same format as the spec since it uses hashmaps instead of tuple's array
